--- BankApp/BankApp/Services/services.swift
@@ -1,11 +1,0 @@
-//
-//  services.swift
-//  BankApp
-//
-//  Created by Joni Campos on 19/09/19.
-//  Copyright © 2019 Joni Campos. All rights reserved.
-//
-
-import Foundation
-
-
